refactor(service/storage): storage provider
